@@ -5,95 +5,60 @@
   "author": "PxFlux Team",
   "private": true,
   "scripts": {
-    "dev": "node build/dev-server.js",
+    "dev": "vite",
     "start": "npm run dev",
-    "build": "node build/build.js",
+    "build": "vite build",
+    "preview": "vite preview",
     "lint": "eslint --ext .js,.vue src"
   },
   "dependencies": {
     "@vimeo/player": "^2.6.1",
     "autosize": "^4.0.2",
-    "axios": "^0.17.1",
+    "axios": "^1.6.0",
     "axios-add-jsonp": "0.0.4",
     "axios-jsonp-pro": "^1.0.1",
-    "element-ui": "^2.2.0",
-    "firebase": "^4.9.1",
-    "firebaseui": "^2.6.0",
+    "element-plus": "^2.4.0",
+    "firebase": "^10.0.0",
+    "firebaseui": "^6.1.0",
     "jsonp": "^0.2.1",
     "keycode": "^2.1.9",
     "latinize": "^0.4.0",
     "medium-editor": "^5.23.3",
     "popper.js": "^1.12.9",
-    "vue": "^2.5.13",
-    "vue-directive-tooltip": "^1.4.2",
-    "vue-input-autowidth": "^1.0.2",
-    "vue-popper-component": "^0.1.0",
-    "vue-popperjs": "^1.2.6",
-    "vue-router": "^3.0.1",
-    "vue-scrollto": "^2.11.0",
-    "vue-select": "^2.4.0",
-    "vue-vimeo-player": "0.0.6",
-    "vue2-dropzone": "^3.0.6",
-    "vue2-medium-editor": "^1.1.6",
-    "vuex": "^3.0.1",
-    "vuex-router-sync": "^5.0.0",
-    "vuexfire": "^2.3.0",
+    "vue": "^3.3.0",
+    "vue-directive-tooltip": "^2.0.0",
+    "vue-input-autowidth": "^2.0.0",
+    "vue-popper-component": "^2.0.0",
+    "vue-router": "^4.2.0",
+    "vue-scrollto": "^2.20.0",
+    "vue-select": "^4.0.0",
+    "vue-vimeo-player": "^1.0.0",
+    "vue3-dropzone": "^2.0.0",
+    "vue3-medium-editor": "^2.0.0",
+    "vuex": "^4.1.0",
+    "vuex-router-sync": "^6.0.0",
     "weakmap": "0.0.6"
   },
   "devDependencies": {
-    "autoprefixer": "^7.2.6",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "chalk": "^2.3.1",
-    "connect-history-api-fallback": "^1.5.0",
-    "copy-webpack-plugin": "^4.4.1",
-    "css-loader": "^0.28.9",
-    "eslint": "^3.19.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-import-resolver-webpack": "^0.8.4",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-html": "^3.0.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-node": "^5.2.1",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.14.1",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.6",
-    "friendly-errors-webpack-plugin": "^1.6.1",
-    "html-webpack-plugin": "^2.30.1",
-    "http-proxy-middleware": "^0.19.1",
-    "lost": "^8.2.0",
+    "@vitejs/plugin-vue": "^4.4.0",
+    "autoprefixer": "^10.4.0",
+    "eslint": "^8.50.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.28.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.1.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "lost": "^9.0.0",
     "megatype": "^1.1.2",
-    "node-sass": "^4.11.0",
-    "opn": "^5.2.0",
-    "optimize-css-assets-webpack-plugin": "^3.2.0",
-    "ora": "^1.4.0",
-    "portfinder": "^1.0.13",
-    "rimraf": "^2.6.0",
-    "sass-loader": "^6.0.6",
-    "semver": "^5.5.0",
-    "shelljs": "^0.7.6",
-    "url-loader": "^1.1.2",
-    "vue-loader": "^13.7.1",
-    "vue-style-loader": "^3.1.2",
-    "vue-template-compiler": "^2.5.13",
-    "webpack": "^3.11.0",
-    "webpack-bundle-analyzer": "^2.10.0",
-    "webpack-dev-middleware": "^1.12.2",
-    "webpack-hot-middleware": "^2.21.0",
-    "webpack-merge": "^4.1.1"
+    "postcss": "^8.4.0",
+    "rimraf": "^5.0.0",
+    "sass": "^1.69.0",
+    "vite": "^4.4.0",
+    "vue-tsc": "^1.8.0"
   },
   "engines": {
-    "node": ">= 4.0.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 16.0.0",
+    "npm": ">= 8.0.0"
   },
   "browserslist": [
     "> 1%",
