@@ -1,18 +1,20 @@
 // The Vue build version to load with the `import` command
 // (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue'
-import { createApp } from './init'
+import { createApp } from 'vue'
+import { createRouter } from './router'
+import { createStore } from './store'
+import { sync } from 'vuex-router-sync'
 import ProgressBar from './components/elements/ProgressBar'
 import firebaseApp from './firebase-app'
 import inputAutoWidth from 'vue-input-autowidth'
 import VueScrollTo from 'vue-scrollto'
+import App from './components/App'
 
 // global progress bar
-const bar = Vue.prototype.$bar = new Vue(ProgressBar).$mount()
-document.body.appendChild(bar.$el)
+let bar = null
 
 // a global mixin that calls `asyncData` when a route component's params change
-Vue.mixin({
+const globalMixin = {
   beforeRouteUpdate (to, from, next) {
     const {asyncData} = this.$options
     if (asyncData) {
@@ -24,12 +26,36 @@ Vue.mixin({
       next()
     }
   }
-})
+}
 
-Vue.use(inputAutoWidth)
-Vue.use(VueScrollTo)
+// Create app instance
+const app = createApp(App)
 
-const {app, router, store} = createApp()
+// Setup router and store
+const router = createRouter()
+const store = createStore()
+
+// Sync router with store
+sync(store, router)
+
+// Install plugins
+app.use(store)
+app.use(router)
+app.use(inputAutoWidth)
+app.use(VueScrollTo)
+
+// Add global mixin
+app.mixin(globalMixin)
+
+// Global progress bar
+bar = createApp(ProgressBar).mount('#progress-bar-placeholder')
+if (!document.getElementById('progress-bar-placeholder')) {
+  const placeholder = document.createElement('div')
+  placeholder.id = 'progress-bar-placeholder'
+  document.body.appendChild(placeholder)
+  bar = createApp(ProgressBar).mount(placeholder)
+}
+app.config.globalProperties.$bar = bar
 
 /**
  * Sync store.state.user with firebase.auth().currentUser
@@ -78,14 +104,18 @@ firebaseApp.auth().onAuthStateChanged(user => {
 })
 
 // wait until router has resolved all async before hooks and async components...
-router.onReady(() => {
+router.isReady().then(() => {
   // Add router hook for handling asyncData.
   // Doing it after initial route is resolved so that we don't double-fetch
   // the data that we already have. Using router.beforeResolve() so that all
   // async components are resolved.
   router.beforeResolve((to, from, next) => {
-    const matched = router.getMatchedComponents(to)
-    const prevMatched = router.getMatchedComponents(from)
+    const matched = to.matched.flatMap(record =>
+      Object.values(record.components).filter(c => c.asyncData)
+    )
+    const prevMatched = from.matched.flatMap(record =>
+      Object.values(record.components).filter(c => c.asyncData)
+    )
     let diffed = false
     const activated = matched.filter((c, i) => {
       return diffed || (diffed = (prevMatched[i] !== c))
@@ -105,7 +135,7 @@ router.onReady(() => {
   })
 
   // actually mount to DOM
-  app.$mount('#app')
+  app.mount('#app')
 })
 
 function b64DecodeUnicode (str) {
